Refactor autotrade endpoints
