+++ api/autotrade/schemas.py
@@ -1,3 +1,4 @@
+from time import time
 from marshmallow import Schema, fields
 from marshmallow.validate import OneOf
