+++ web/.env
@@ -28,7 +28,7 @@ REACT_APP_ACCOUNT_BALANCE_GBP=/account/gbp
 REACT_APP_ACCOUNT_BALANCE_RAW=/account/balance/raw
 REACT_APP_TICKER=/account/ticker
 REACT_APP_RESEARCH_BLACKLIST=/research/blacklist
-REACT_APP_RESEARCH_CONTROLLER=/research/controller
+REACT_APP_RESEARCH_CONTROLLER=/autotrade-settings/bots
 REACT_APP_NO_CANNIBALISM_SYMBOLS=/account/symbols/no-cannibal
 REACT_APP_BALANCE_ESTIMATE=/account/balance/estimate
 REACT_APP_BINANCE_INFO="https://api.binance.com/api/v3/exchangeInfo"
@@ -38,4 +38,4 @@ REACT_APP_TEST_BOT=/paper-trading
 REACT_APP_ARCHIVE_TEST_BOT=/paper-trading/archive
 REACT_APP_ACTIVATE_TEST_BOT=/paper-trading/activate
 REACT_APP_DEACTIVATE_TEST_BOT=/paper-trading/deactivate
-REACT_APP_TEST_AUTOTRADE_SETTINGS=/research/test-autotrade-settings
+REACT_APP_TEST_AUTOTRADE_SETTINGS=/autotrade-settings/paper-trading
